Tests/LibWeb/Text/expected/css/style-invalidation/sheet-add-hover-pseudo-only.txt

:hover-only sheet add: fullStyleInvalidations=1, styleInvalidations=3, elementStyleRecomputations=33, elementStyleNoopRecomputations=32
